moved load user logic to header component
