+++ src/screens/Dashboard/index.tsx
@@ -10,17 +10,6 @@ import { Container, ContentText } from "./styles";
 
 export default function Dashboard() {
   const { signOut, session } = useAuth();
-  const [user, setUser] = useState<{
-    id?: string;
-    full_name: string;
-    avatar_url: string;
-  }>({
-    id: "",
-    full_name: "",
-    //TODO: use a placeholder avatar
-    avatar_url:
-      "https://kmqurfaofmowoonastqb.supabase.co/storage/v1/object/sign/avatars/avatar1.png?token=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1cmwiOiJhdmF0YXJzL2F2YXRhcjEucG5nIiwiaWF0IjoxNjU0OTgwNDU2LCJleHAiOjE5NzAzNDA0NTZ9.bYJuCu2pk5Eeug9T8p8ZZYOKX_pZZhY9coZgOqPPiCs",
-  });
 
   //TODO: move this logic to useEffect
   async function handleGetUserGroups() {
@@ -29,28 +18,7 @@ export default function Dashboard() {
       const { data, error } = await supabase
         .from("user_groups")
         .select("*,group_id(name)")
-        .eq("user_id", user?.id);
-
-      console.log({ data });
-    } catch (error) {
-      if (error instanceof AppError) return Alert.alert(error.message);
-      console.log(`unknown ERROR: ${error}`);
-    }
-  }
-
-  async function loadUser() {
-    try {
-      const { data, error } = await supabase
-        .from("profiles")
-        .select("id,full_name,avatar_url")
-        .eq("id", session?.user?.id);
-      console.log(data);
-      if (!error && data && session)
-        setUser({
-          id: session?.user?.id,
-          full_name: data[0].full_name,
-          avatar_url: data[0].avatar_url,
-        });
+        .eq("user_id", session?.user?.id);
     } catch (error) {
       if (error instanceof AppError) return Alert.alert(error.message);
       console.log(`unknown ERROR: ${error}`);
@@ -66,19 +34,11 @@ export default function Dashboard() {
     }
   }
 
-  useEffect(() => {
-    loadUser();
-  }, []);
-
   return (
     <Container>
-      <Header
-        name={user.full_name}
-        logoutFunction={handleSignOut}
-        avatarUri={user!.avatar_url}
-      />
-      <ContentText>{user!.full_name}</ContentText>
-      <Button title="Logout" onPress={loadUser} />
+      <Header userId={session?.user?.id} logoutFunction={handleSignOut} />
+      <ContentText>{session?.user?.email}</ContentText>
+      <Button title="Logout" onPress={handleSignOut} />
     </Container>
   );
 }
